feat: unread message content types

+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/data/conversation/Conversation.kt
@@ -128,7 +128,7 @@ sealed class ConversationDetails(open val conversation: Conversation) {
         val otherUser: OtherUser,
         val legalHoldStatus: LegalHoldStatus,
         val userType: UserType,
-        val unreadMessagesCount: Long = 0L,
+        val unreadMessagesCount: Int = 0,
         val unreadMentionsCount: Long = 0L,
         val lastUnreadMessage: Message?
     ) : ConversationDetails(conversation)
@@ -137,7 +137,7 @@ sealed class ConversationDetails(open val conversation: Conversation) {
         override val conversation: Conversation,
         val legalHoldStatus: LegalHoldStatus,
         val hasOngoingCall: Boolean = false,
-        val unreadMessagesCount: Long = 0L,
+        val unreadMessagesCount: Int = 0,
         val unreadMentionsCount: Long = 0L,
         val lastUnreadMessage: Message?,
         val isSelfUserMember: Boolean,
